\name{qqhist_y}
\alias{qqhist_y}
\alias{ranefqqhist}
\title{
Histograms and qqplots for observed residuals and random effects
}
\description{
Histograms and qqplots for observed residuals and random effects estimates as available by using ranef, lmer
}
\usage{
qqhist_y(resadapintervals, outertitle = "")

ranefqqhist(resadapintervals, outertitle = "")
}

\arguments{
  \item{resadapintervals}{ an object as can be obtained by \code{\link{adaplmmintervals}}

}
  \item{outertitle}{ a cahracter string, as an overall title of the plots

}
}

\keyword{ hplot }